Support touch events for the range slider.
Simply reuse the same handlers as for the mouse.

I use Chromatix as a dedicated full screen media player on raspberry Pis
with a touchscreen for my kids, and not being able to skip back and forward
for audiobooks/stories can be frustrating.

This has been the only complaint so far :)